macOS: bake GeneralsX DXVK patchset into source

+++ src/util/util_small_vector.h
@@ -194,7 +194,7 @@ namespace dxvk {
 
     size_t pick_capacity(size_t n) {
       // Pick next largest power of two for the new capacity
-      return size_t(1u) << ((sizeof(n) * 8u) - bit::lzcnt(n - 1));
+      return size_t(1u) << ((sizeof(n) * 8u) - bit::lzcnt(static_cast<uint64_t>(n - 1)));
     }
 
     T* ptr(size_t idx) {
